added num of hidden layers

+++ python/gps/algorithm/algorithm_mdgps.py
@@ -239,9 +239,9 @@ class AlgorithmMDGPS(Algorithm):
         """
         # Compute previous cost and previous expected cost.
         prev_M = len(self.prev) # May be different in future.
-        prev_laplace = np.arange(prev_M)
-        prev_mc = np.arange(prev_M)
-        prev_predicted = np.arange(prev_M)
+        prev_laplace = np.arange(prev_M).astype(np.float32)
+        prev_mc = np.arange(prev_M).astype(np.float32)
+        prev_predicted = np.arange(prev_M).astype(np.float32)
         for m in range(prev_M):
             prev_nn = self.prev[m].pol_info.traj_distr()
             prev_lg = self.prev[m].new_traj_distr
@@ -265,8 +265,8 @@ class AlgorithmMDGPS(Algorithm):
             ).sum()
 
         # Compute current cost.
-        cur_laplace = np.arange(self.M)
-        cur_mc = np.arange(self.M)
+        cur_laplace = np.arange(self.M).astype(np.float32)
+        cur_mc = np.arange(self.M).astype(np.float32)
         for m in range(self.M):
             if self._hyperparams['ioc']:
                 self._eval_cost(m, prev_cost=True)
